requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html 
torch==1.9.0+cu111
-f https://download.pytorch.org/whl/torch_stable.html 
torchvision==0.10.0
torchaudio==0.9.0
pytorch-lightning==1.4.2
nnAudio==0.2.6
AudioLoader==0.0.3
fastwer
pandas
tabulate
mido
joblib
IPython
matplotlib
omegaconf
hydra-core==1.1.1
mir_eval
scikit-learn